+++ src/ProxyClient/Http/HttpAdapter.php
@@ -14,11 +14,14 @@ namespace FOS\HttpCache\ProxyClient\Http;
 use FOS\HttpCache\Exception\ExceptionCollection;
 use FOS\HttpCache\Exception\InvalidArgumentException;
 use FOS\HttpCache\Exception\InvalidUrlException;
+use FOS\HttpCache\Exception\MissingHostException;
 use FOS\HttpCache\Exception\ProxyResponseException;
 use FOS\HttpCache\Exception\ProxyUnreachableException;
 use Http\Client\Exception\HttpException;
 use Http\Client\Exception\RequestException;
 use Http\Client\HttpAsyncClient;
+use Http\Discovery\HttpAsyncClientDiscovery;
+use Http\Discovery\UriFactoryDiscovery;
 use Http\Message\UriFactory;
 use Http\Promise\Promise;
 use Psr\Http\Message\RequestInterface;
@@ -63,21 +66,27 @@ class HttpAdapter
     private $baseUri;
 
     /**
-     * @param string[]        $servers    Caching proxy server hostnames or IP
-     *                                    addresses, including port if not port 80.
-     *                                    E.g. ['127.0.0.1:6081']
-     * @param string          $baseUri    Default application hostname, optionally
-     *                                    including base URL, for purge and refresh
-     *                                    requests (optional). This is required if
-     *                                    you purge and refresh paths instead of
-     *                                    absolute URLs
-     * @param HttpAsyncClient $httpClient
-     * @param UriFactory      $uriFactory
+     * @param string[]             $servers    Caching proxy server hostnames or IP
+     *                                         addresses, including port if not port 80.
+     *                                         E.g. ['127.0.0.1:6081']
+     * @param string               $baseUri    Default application hostname, optionally
+     *                                         including base URL, for purge and refresh
+     *                                         requests (optional). This is required if
+     *                                         you purge and refresh paths instead of
+     *                                         absolute URLs
+     * @param HttpAsyncClient|null $httpClient Client capable of sending HTTP requests. If no
+     *                                         client is supplied, a default one is created
+     * @param UriFactory|null      $uriFactory Factory for PSR-7 URIs. If not specified, a
+     *                                         default one is created
      */
-    public function __construct(array $servers, $baseUri, HttpAsyncClient $httpClient, UriFactory $uriFactory)
-    {
-        $this->httpClient = $httpClient;
-        $this->uriFactory = $uriFactory;
+    public function __construct(
+        array $servers,
+        $baseUri = '',
+        HttpAsyncClient $httpClient = null,
+        UriFactory $uriFactory = null
+    ) {
+        $this->httpClient = $httpClient ?: HttpAsyncClientDiscovery::find();
+        $this->uriFactory = $uriFactory ?: UriFactoryDiscovery::find();
 
         $this->setServers($servers);
         $this->setBaseUri($baseUri);
@@ -90,6 +99,13 @@ class HttpAdapter
      */
     public function invalidate(RequestInterface $invalidationRequest)
     {
+        if (!$this->baseUri && !$invalidationRequest->getUri()->getHost()) {
+            throw new MissingHostException(sprintf(
+                'URI "%s" is not absolute. Either configure the base URI or invalidate with absolute URLs including the host.',
+                (string) $invalidationRequest->getUri()
+            ));
+        }
+
         $signature = $this->getRequestSignature($invalidationRequest);
 
         if (isset($this->queue[$signature])) {
@@ -229,7 +245,7 @@ class HttpAdapter
      */
     private function setBaseUri($uriString = null)
     {
-        if (null === $uriString) {
+        if (!$uriString) {
             $this->baseUri = null;
 
             return;
@@ -254,6 +270,13 @@ class HttpAdapter
      */
     private function filterUri($uriString, array $allowedParts = [])
     {
+        if (!is_string($uriString)) {
+            throw new \InvalidArgumentException(sprintf(
+                'Uri parameter must be a string, %s given',
+                gettype($uriString)
+            ));
+        }
+
         // Creating a PSR-7 URI without scheme (with parse_url) results in the
         // original hostname to be seen as path. So first add a scheme if none
         // is given.