+++ tests/Unit/ProxyClient/HttpAdapterTest.php
@@ -12,17 +12,17 @@
 namespace FOS\HttpCache\Tests\Unit\ProxyClient;
 
 use FOS\HttpCache\Exception\ExceptionCollection;
-use FOS\HttpCache\ProxyClient\Varnish;
+use FOS\HttpCache\ProxyClient\Http\HttpAdapter;
+use Http\Message\MessageFactory;
 use Http\Mock\Client;
 use Http\Client\Exception\RequestException;
 use Http\Discovery\MessageFactoryDiscovery;
-use Http\Promise\Promise;
 use Psr\Http\Message\RequestInterface;
 
 /**
- * Testing the base methods of the proxy client, using the Varnish client as concrete class.
+ * Testing the HTTP Adapter
  */
-class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
+class HttpAdaptertTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * Mock client.
@@ -30,6 +30,17 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
      * @var Client
      */
     private $client;
+
+    /**
+     * @var MessageFactory
+     */
+    private $messageFactory;
+
+    protected function setUp()
+    {
+        $this->client = new Client();
+        $this->messageFactory = MessageFactoryDiscovery::find();
+    }
 
     /**
      * @dataProvider exceptionProvider
@@ -41,12 +52,15 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
     public function testExceptions(\Exception $exception, $type, $message = null)
     {
         $this->client->addException($exception);
-        $varnish = new Varnish(['127.0.0.1:123'], ['base_uri' => 'my_hostname.dev'], $this->client);
-
-        $varnish->purge('/');
+        $httpAdapter = new HttpAdapter(
+            ['127.0.0.1:123'],
+            'my_hostname.dev',
+            $this->client
+        );
+        $httpAdapter->invalidate($this->messageFactory->createRequest('PURGE', '/path'));
 
         try {
-            $varnish->flush();
+            $httpAdapter->flush();
             $this->fail('Should have aborted with an exception');
         } catch (ExceptionCollection $exceptions) {
             $this->assertCount(1, $exceptions);
@@ -60,7 +74,8 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
         }
 
         // Queue must now be empty, so exception above must not be thrown again.
-        $varnish->purge('/path')->flush();
+        $httpAdapter->invalidate($this->messageFactory->createRequest('GET', '/path'));
+        $httpAdapter->flush();
     }
 
     public function exceptionProvider()
@@ -87,15 +102,16 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
     {
         $this->markTestSkipped('Default php-http behaviour is not to exceptions');
 
-        $response = MessageFactoryDiscovery::find()->createResponse(
+        $response = $this->messageFactory->createResponse(
             405,
             'Not allowed'
         );
         $this->client->addResponse($response);
 
-        $varnish = new Varnish(['127.0.0.1:123'], ['base_uri' => 'my_hostname.dev'], $this->client);
+        $httpAdapter = new HttpAdapter(['127.0.0.1:123'], ['base_uri' => 'my_hostname.dev'], $this->client);
+        $httpAdapter->invalidate($this->messageFactory->createRequest('PURGE', '/'));
         try {
-            $varnish->purge('/')->flush();
+            $httpAdapter->flush();
             $this->fail('Should have aborted with an exception');
         } catch (ExceptionCollection $exceptions) {
             $this->assertCount(1, $exceptions);
@@ -108,26 +124,47 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
 
     /**
      * @expectedException \FOS\HttpCache\Exception\MissingHostException
-     * @expectedExceptionMessage cannot be invalidated without a host
+     * @expectedExceptionMessage is not absolute
      */
     public function testMissingHostExceptionIsThrown()
     {
-        $varnish = new Varnish(['127.0.0.1:123'], [], $this->client);
-        $varnish->purge('/path/without/hostname');
+        $httpAdapter = new HttpAdapter(
+            ['127.0.0.1:123'],
+            '',
+            $this->client
+        );
+
+        $request = $this->messageFactory->createRequest('PURGE', '/path/without/hostname');
+        $httpAdapter->invalidate($request);
     }
 
     public function testSetBasePathWithHost()
     {
-        $varnish = new Varnish(['127.0.0.1'], ['base_uri' => 'fos.lo'], $this->client);
-        $varnish->purge('/path')->flush();
+        $httpAdapter = new HttpAdapter(
+            ['127.0.0.1'],
+            'fos.lo',
+            $this->client
+        );
+
+        $request = $this->messageFactory->createRequest('PURGE', '/path');
+        $httpAdapter->invalidate($request);
+        $httpAdapter->flush();
+
         $requests = $this->getRequests();
         $this->assertEquals('fos.lo', $requests[0]->getHeaderLine('Host'));
     }
 
     public function testSetBasePathWithPath()
     {
-        $varnish = new Varnish(['127.0.0.1'], ['base_uri' => 'http://fos.lo/my/path'], $this->client);
-        $varnish->purge('append')->flush();
+        $httpAdapter = new HttpAdapter(
+            ['127.0.0.1'],
+            'http://fos.lo/my/path',
+            $this->client
+        );
+        $request = $this->messageFactory->createRequest('PURGE', 'append');
+        $httpAdapter->invalidate($request);
+        $httpAdapter->flush();
+
         $requests = $this->getRequests();
         $this->assertEquals('fos.lo', $requests[0]->getHeaderLine('Host'));
         $this->assertEquals('http://127.0.0.1/my/path/append', (string) $requests[0]->getUri());
@@ -135,8 +172,11 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
 
     public function testSetServersDefaultSchemeIsAdded()
     {
-        $varnish = new Varnish(['127.0.0.1'], ['base_uri' => 'fos.lo'], $this->client);
-        $varnish->purge('/some/path')->flush();
+        $httpAdapter = new HttpAdapter(['127.0.0.1'], 'fos.lo', $this->client);
+        $request = $this->messageFactory->createRequest('PURGE', '/some/path');
+        $httpAdapter->invalidate($request);
+        $httpAdapter->flush();
+
         $requests = $this->getRequests();
         $this->assertEquals('http://127.0.0.1/some/path', $requests[0]->getUri());
     }
@@ -147,7 +187,7 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetServersThrowsInvalidUrlException()
     {
-        new Varnish(['http:///this is no url']);
+        new HttpAdapter(['http:///this is no url']);
     }
 
     /**
@@ -156,22 +196,24 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetServersThrowsWeirdInvalidUrlException()
     {
-        new Varnish(['this ://is no url']);
+        new HttpAdapter(['this ://is no url']);
     }
 
     /**
      * @expectedException \FOS\HttpCache\Exception\InvalidUrlException
-     * @expectedExceptionMessage Server "http://127.0.0.1:80/some/weird/path" is invalid. Only scheme, host, port URL parts are allowed
+     * @expectedExceptionMessage Server "http://127.0.0.1:80/some/path" is invalid. Only scheme, host, port URL parts are allowed
      */
     public function testSetServersThrowsInvalidServerException()
     {
-        new Varnish(['http://127.0.0.1:80/some/weird/path']);
+        new HttpAdapter(['http://127.0.0.1:80/some/path']);
     }
 
     public function testFlushEmpty()
     {
-        $varnish = new Varnish(['127.0.0.1', '127.0.0.2'], ['base_uri' => 'fos.lo'], $this->client);
-        $this->assertEquals(0, $varnish->flush());
+        $httpAdapter = new HttpAdapter(
+            ['127.0.0.1', '127.0.0.2'], 'fos.lo', $this->client
+        );
+        $this->assertEquals(0, $httpAdapter->flush());
 
         $this->assertCount(0, $this->client->getRequests());
     }
@@ -202,14 +244,17 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
             )
             ->getMock();
 
-        $varnish = new Varnish(['127.0.0.1', '127.0.0.2'], ['base_uri' => 'fos.lo'], $httpClient);
+        $httpAdapter = new HttpAdapter(
+            ['127.0.0.1', '127.0.0.2'],
+            'fos.lo',
+            $httpClient
+        );
+        $httpAdapter->invalidate($this->messageFactory->createRequest('PURGE', '/a'));
+        $httpAdapter->invalidate($this->messageFactory->createRequest('PURGE', '/b'));
 
         $this->assertEquals(
             2,
-            $varnish
-                ->purge('/c')
-                ->purge('/b')
-                ->flush()
+            $httpAdapter->flush()
         );
     }
 
@@ -235,22 +280,22 @@ class AbstractProxyClientTest extends \PHPUnit_Framework_TestCase
             )
             ->getMock();
 
-        $varnish = new Varnish(['127.0.0.1', '127.0.0.2'], ['base_uri' => 'fos.lo'], $httpClient);
+        $httpAdapter = new HttpAdapter(['127.0.0.1', '127.0.0.2'], 'fos.lo', $httpClient);
+        $httpAdapter->invalidate(
+            $this->messageFactory->createRequest('PURGE', '/c', ['a' => 'b', 'c' => 'd'])
+        );
+        $httpAdapter->invalidate(
+            // same request (header order is not significant)
+            $this->messageFactory->createRequest('PURGE', '/c', ['c' => 'd', 'a' => 'b'])
+        );
+        // different request as headers different
+        $httpAdapter->invalidate($this->messageFactory->createRequest('PURGE', '/c'));
+        $httpAdapter->invalidate($this->messageFactory->createRequest('PURGE', '/c'));
 
         $this->assertEquals(
             2,
-            $varnish
-                ->purge('/c', ['a' => 'b', 'c' => 'd'])
-                ->purge('/c', ['c' => 'd', 'a' => 'b']) // same request (header order is not significant)
-                ->purge('/c') // different request as headers different
-                ->purge('/c')
-                ->flush()
+            $httpAdapter->flush()
         );
-    }
-
-    protected function setUp()
-    {
-        $this->client = new Client();
     }
 
     /**