+++ tests/Unit/ResponseTaggerTest.php
@@ -11,6 +11,7 @@
 
 namespace FOS\HttpCache\Tests\Unit;
 
+use FOS\HttpCache\ProxyClient\Http\HttpAdapter;
 use FOS\HttpCache\ResponseTagger;
 use FOS\HttpCache\ProxyClient\Varnish;
 use FOS\HttpCache\Exception\InvalidTagException;
@@ -58,7 +59,8 @@ class ResponseTaggerTest extends \PHPUnit_Framework_TestCase
 
     public function testStrictEmptyTag()
     {
-        $proxyClient = new Varnish(['localhost']);
+        $httpAdapter = new HttpAdapter(['localhost'], 'localhost');
+        $proxyClient = new Varnish($httpAdapter);
 
         $tagHandler = new ResponseTagger($proxyClient, array('strict' => true));